refactor: 피드백 반영 Repository 묶어 주기
