--- src/main/java/repository/Repository.java
@@ -1,4 +1,0 @@
-package repository;
-
-public interface Repository<T> {
-}
